Rename pascal cased files
