+++ public_src/js/bpmn/resource/resource.json
