+++ public_src/js/bpmn/resource/resourceContextPad.js
