+++ public_src/js/bpmn/resource/resourceElementFactory.js
