+++ public_src/js/bpmn/resource/resourceRenderer.js
